+++ src/Marsen.NetCore.Dojo/Kata_ShowHands/OnePair.cs
@@ -1,0 +1,17 @@
+﻿using System.Collections.Generic;
+using System.Linq;
+
+namespace Marsen.NetCore.Dojo.Kata_ShowHands
+{
+    public class OnePair : ICategoryRule
+    {
+        public bool Apply(List<Card> cardList)
+        {
+            return cardList
+                       .GroupBy(x => x.Rank)
+                       .Select(g => new {Count = g.Count(), Rank = g.Key}).Count(x => x.Count == 2) == 1;
+        }
+
+        public Category Category => Category.OnePair;
+    }
+}
